[Ariawan] Update integration test rent  cars with jest

+++ __tests__/Integration/v1/cars/:id/rentCars.test.js
@@ -21,6 +21,8 @@ describe("POST /v1/cars/:id/rent", () => {
 
   let idCarsRent = 0;
 
+  const idCarNotFound = 1000;
+
   // password admin and customer
   const password = "Hati-hati-dijalan";
   // data admin for authentication to get token admin
@@ -31,41 +33,45 @@ describe("POST /v1/cars/:id/rent", () => {
     roleId: 2, // admin role 2
   };
   // data customer for authentication to get token customer
-  const userRentCustomer = {
+  const userRentCustomerRent = {
     name: "Customer Rent Test",
-    email: "customerrent@mail.test",
+    email: "customerrent@gmail.com",
     encryptedPassword: bcrypt.hashSync(password, 10), // hash of password
     roleId: 1, // admin role 1
   };
   beforeEach(async () => {
     // create user admin and customer and crate car before it
     await User.create(userRentAdmin);
-    await User.create(userRentCustomer);
+    await User.create(userRentCustomerRent);
     const addCar = await Car.create(carDataRent);
     idCarsRent = addCar.id;
+    const findNotFoundCars = await Car.findByPk(idCarNotFound);
+    if (findNotFoundCars) {
+      await Car.destroy({
+        where: {
+          id: idCarNotFound,
+        },
+      });
+    }
   });
 
   afterEach(async () => {
     // delete user admin and customer after it
-    try {
-      await User.destroy({
-        where: {
-          email: userRentAdmin.email,
-        },
-      });
-      await User.destroy({
-        where: {
-          email: userRentCustomer.email,
-        },
-      });
-      await Car.destroy({
-        where: {
-          id: idCarsRent,
-        },
-      });
-    } catch (err) {
-      console.error(err.message); // error message
-    }
+    await User.destroy({
+      where: {
+        email: userRentAdmin.email.toLowerCase(),
+      },
+    });
+    // await User.destroy({
+    //   where: {
+    //     email: userRentCustomerRent.email.toLowerCase(),
+    //   },
+    // });
+    // await Car.destroy({
+    //   where: {
+    //     id: idCarsRent,
+    //   },
+    // });
   });
 
   it("should response with 401 as status code (userAdmin cannot be update car)", async () => {
@@ -90,6 +96,64 @@ describe("POST /v1/cars/:id/rent", () => {
                 },
               },
             });
+          });
+      });
+  });
+  it("should response with 404 as status code (Cars not found for rent)", async () => {
+    return await request(app)
+      .post("/v1/auth/login") // request api login
+      .set("Content-Type", "application/json")
+      .send({ email: userRentCustomerRent.email, password: password }) // need email and password for login userCustomer
+      .then(async (res) => {
+        await request(app)
+          .post("/v1/cars/" + idCarNotFound + "/rent") // request api create cars
+          .set("authorization", "Bearer " + res.body.accessToken) // set authorization jwt
+          .send(rentData)
+          .then(async (notFoundCarsRent) => {
+            await expect(notFoundCarsRent.statusCode).toBe(404); // check status respond
+            await expect(notFoundCarsRent.body.error.message).toEqual(
+              "Not found car!"
+            );
+          });
+      });
+  });
+  it("should response with 201 as status code (Suksess Rent Cars)", async () => {
+    return await request(app)
+      .post("/v1/auth/login") // request api login
+      .set("Content-Type", "application/json")
+      .send({ email: userRentCustomerRent.email, password: password }) // need email and password for login userCustomer
+      .then(async (res) => {
+        await request(app)
+          .post("/v1/cars/" + idCarsRent + "/rent") // request api create cars
+          .set("authorization", "Bearer " + res.body.accessToken) // set authorization jwt
+          .send(rentData)
+          .then(async (notFoundCarsRent) => {
+            await expect(notFoundCarsRent.statusCode).toBe(201); // check status respond
+            await expect(notFoundCarsRent.body.carId).toBe(idCarsRent); // check status respond
+          });
+      });
+  });
+  it("should response with 422 as status code (al)", async () => {
+    return await request(app)
+      .post("/v1/auth/login") // request api login
+      .set("Content-Type", "application/json")
+      .send({ email: userRentCustomerRent.email, password: password }) // need email and password for login userCustomer
+      .then(async (res) => {
+        await request(app)
+          .post("/v1/cars/" + idCarsRent + "/rent") // request api create cars
+          .set("authorization", "Bearer " + res.body.accessToken) // set authorization jwt
+          .send(rentData)
+          .then(async () => {
+            await request(app)
+              .post("/v1/cars/" + idCarsRent + "/rent") // request api create cars
+              .set("authorization", "Bearer " + res.body.accessToken) // set authorization jwt
+              .send(rentData)
+              .then(async (alreadyRent) => {
+                await expect(alreadyRent.statusCode).toBe(422); // check status respond
+                await expect(alreadyRent.body.error.message).toEqual(
+                  "Car is already rented!!"
+                );
+              });
           });
       });
   });
